Rearrange all the things
